[feat] 로컬에 저장된 토큰을 불러오는 코드 단순화 & 토큰 갱신 로직 개발 (HH-221)

+++ lib/data/remote/repository/comment_repository.dart
@@ -1,17 +1,15 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:http/http.dart' as http;
 import 'package:pocket_pose/config/api_url.dart';
 import 'package:pocket_pose/data/entity/response/comment_list_response.dart';
+import 'package:pocket_pose/data/remote/provider/kakao_login_provider.dart';
 import 'package:pocket_pose/domain/entity/comment_data.dart';
 
-const _storage = FlutterSecureStorage();
-const _accessTokenKey = 'kakaoAccessToken';
-const _refreshTokenKey = 'kakaoRefreshToken';
-
 class CommentRepository {
+  KaKaoLoginProvider loginProvider = KaKaoLoginProvider();
+
   Future<CommentListResponse> getComments(String videoId) async {
     final url = Uri.parse('${AppUrl.commentUrl}/$videoId');
 
@@ -41,8 +39,10 @@ class CommentRepository {
   Future<bool> postComment(String videoId, String content) async {
     final url = Uri.parse('${AppUrl.commentUrl}/$videoId');
 
-    final accessToken = await _storage.read(key: _accessTokenKey);
-    final refreshToken = await _storage.read(key: _refreshTokenKey);
+    await loginProvider.checkAccessToken();
+
+    final accessToken = loginProvider.accessToken;
+    final refreshToken = loginProvider.refreshToken;
 
     final headers = <String, String>{
       'Content-Type': 'application/json;charset=UTF-8',
@@ -60,6 +60,9 @@ class CommentRepository {
 
     if (response.statusCode == 200) {
       debugPrint("댓글 등록 성공! json: $json");
+
+      loginProvider.updateToken(response.headers);
+
       return true;
     } else {
       debugPrint('댓글 등록 실패 json $json');
@@ -70,12 +73,15 @@ class CommentRepository {
   Future<bool> deleteComment(String commentId) async {
     final url = Uri.parse('${AppUrl.commentUrl}/$commentId');
 
-    //final accessToken = await _storage.read(key: _accessTokenKey);
-    final refreshToken = await _storage.read(key: _refreshTokenKey);
+    await loginProvider.checkAccessToken();
+
+    final accessToken = loginProvider.accessToken;
+    final refreshToken = loginProvider.refreshToken;
 
     final headers = <String, String>{
       'Content-Type': 'application/json;charset=UTF-8',
-      if (refreshToken != null) "cookie": "x-refresh-token=$refreshToken"
+      if (accessToken != null && refreshToken != null)
+        "cookie": "x-access-token=$accessToken;x-refresh-token=$refreshToken"
     };
 
     final response = await http.delete(url, headers: headers);
@@ -84,16 +90,7 @@ class CommentRepository {
     if (response.statusCode == 200) {
       debugPrint("댓글 삭제 성공! json: $json");
 
-      final headers = response.headers;
-      final cookies = response.headers['cookie'];
-      // final accessToken = loginProvider.extractToken(cookies, 'x-access-token');
-      // final refreshToken =
-      //     loginProvider.extractToken(cookies, 'x-refresh-token');
-
-      debugPrint("댓글 목록 headers: $headers");
-      debugPrint("댓글 목록 cookies: $cookies");
-      // debugPrint("댓글 목록 accessToken: $accessToken");
-      // debugPrint("댓글 목록 refreshToken: $refreshToken");
+      loginProvider.updateToken(response.headers);
 
       return true;
     } else {
